requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21
Pillow>=9.0
scikit-learn>=1.0
pandas>=1.5
matplotlib
seaborn
tqdm
PyYAML
h5py
typing-extensions
conch @ git+https://github.com/Mahmoodlab/CONCH.git@02d6ac59cc20874bff0f581de258c2b257f69a84#egg=conch
openai>=1.0.0
requests
filelock
regex
timm
fsspec
huggingface_hub
transformers
tokenizers